+++ gui/combat_screen.py
@@ -9,6 +9,7 @@ import gui.gui_constants as constants
 from gui.asset_cache import AssetCache
 from gui.card import Card
 from gui.tooltips import Tooltip
+from gui.combat_log import CombatLog
 
 class Hand(FloatLayout):
     """Widget representing the player's hand of cards."""
@@ -59,50 +60,6 @@ class Hand(FloatLayout):
         for card in self.children[:]:
             self.remove_widget(card)
         self.position_cards()
-
-
-class CombatLog(Widget):
-    """Widget representing the combat log."""
-    combat_log_label = ObjectProperty(None)
-
-    def __init__(self, **kwargs):
-        """Initialize a new CombatLog."""
-        super().__init__(**kwargs)
-        self.history = []
-        self.pending = []
-        self.log_shown = False
-
-    def flush_log_messages(self, event_manager):
-        """Writes all pending log messages to the log display."""
-        self.pending += event_manager.logger.get_combat_logs()
-        if self.pending:
-            Clock.schedule_interval(self.log_message, 0.33)
-    
-    def log_message(self, dt):
-        if self.pending:
-            message = self.pending.pop(0)
-            self.history.append(message)
-            if self.combat_log_label.text:
-                self.combat_log_label.text += "\n"
-            self.combat_log_label.text += message
-        if not self.pending:
-            Clock.unschedule(self.log_message)
-            if not self.log_shown:
-                Clock.schedule_once(self.hide_message, 2)  
-    
-    def hide_message(self, dt):
-        """Hides the current message after a delay."""
-        self.combat_log_label.text = ""
-
-    def show_history(self):
-        """Displays the combat log history."""
-        self.combat_log_label.text = "\n".join(self.history)
-        self.log_shown = True
-    
-    def hide_history(self):
-        """Hides the combat log history."""
-        self.combat_log_label.text = ""
-        self.log_shown = False
 
 
 class ScreenDarken(Widget):
